Minor Inventory Panel file reorganization
