+++ app/layout/project/sidebar/project/inventory-panel/constants.ts
@@ -1,12 +1,12 @@
 import { NavigationInventoryTabs } from 'layout/project/navigation/types';
 
+import CostSurfaceTable from './cost-surface';
 import CostSurfaceInfo from './cost-surface/info';
-import CostSurfaceTable from './cost-surface/table';
+import FeaturesTable from './features';
 import FeaturesInfo from './features/info';
 import FeatureUploadModal from './features/modals/upload';
-import FeaturesTable from './features/table';
+import ProtectedAreasTable from './protected-areas';
 import ProtectedAreasFooter from './protected-areas/footer';
-import ProtectedAreasTable from './protected-areas/table';
 import { Panel } from './types';
 
 export const INVENTORY_TABS = {
