+++ app/layout/project/sidebar/project/inventory-panel/cost-surface/index.tsx
@@ -1,7 +1,5 @@
-import Section from 'layout/section';
-
-const InventoryPanelCostSurface = (): JSX.Element => {
-  return <Section>InventoryPanelCostSurface</Section>;
+const InventoryPanelCostSurface = ({ noData: noDataMessage }: { noData: string }): JSX.Element => {
+  return <div className="flex h-[200px] items-center justify-center">{noDataMessage}</div>;
 };
 
 export default InventoryPanelCostSurface;
