--- app/layout/project/sidebar/project/inventory-panel/cost-surface/table/index.tsx
@@ -1,5 +1,0 @@
-const CostSurfaceTable = ({ noData: noDataMessage }: { noData: string }): JSX.Element => {
-  return <div className="flex h-[200px] items-center justify-center">{noDataMessage}</div>;
-};
-
-export default CostSurfaceTable;
