+++ app/layout/project/sidebar/project/inventory-panel/protected-areas/index.tsx
@@ -1,7 +1,9 @@
-import Section from 'layout/section';
-
-const InventoryPanelProtectedAreas = (): JSX.Element => {
-  return <Section>InventoryPanelProtectedAreas</Section>;
+const InventoryPanelProtectedAreas = ({
+  noData: noDataMessage,
+}: {
+  noData: string;
+}): JSX.Element => {
+  return <div className="flex h-[200px] items-center justify-center">{noDataMessage}</div>;
 };
 
 export default InventoryPanelProtectedAreas;
